requirements.txt
azure-eventhub==5.3.1
Flask==1.1.2
twilio==6.54.0